requirements.txt
python-dotenv
requests
schedule
tqdm
